Changed prices to use doubles.

+++ ShopGUI-Copy/src/me/bukkit/fujinuji/gui/PriceManager.java
@@ -75,7 +75,7 @@ public class PriceManager
     player.openInventory(inventory);
   }
   
-  public static void openPriceSet(String action, Player player, int price)
+  public static void openPriceSet(String action, Player player, double price)
   {
     if (price < 0)
     {
